getSpecific getAll addObject updateObject deleteObject
the following methods have been added for all objects with proper copies inplace such that we aren't giving out the actual object/list only a copy of it.

Furthermore, I refactored the code such that instead of having to maually copy each attribute of an object, I made it such that with a createCopy function you only need to add an id and the object you wish to copy.

All the objects follow a similar structure thus if an issue exists in one, it will likely be in the others.

For the following object classes, I needed to add an ID variable such that it would work properly with the HashMap stub database I have.
- Appointment
- MedicalHistory
- PatientVisitSummary